refactor: split Backup dialog
