+++ packages/maskbook/src/extension/options-page/DashboardDialogs/Backup/index.tsx
@@ -1,0 +1,2 @@
+export { DashboardBackupDialog } from './Backup'
+export { DashboardRestoreDialog } from './Restore'
